backend/requirements.txt
# Django Framework and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
django-filter==23.5 

# SimpleJWT for Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.9.0

# Database and ORM Support
psycopg2-binary==2.9.9
sqlparse==0.5.1

# Other Utilities
asgiref==3.8.1
Cython==3.0.10
pillow==10.4.0
tzdata==2024.1
whitenoise

# Web Server
gunicorn==21.2.0
